+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Upgraded to NativeScript 7.0.
 - Do not add second linebreak at the end of file.
+- Added support for `indent` fava option.
 
 ## 1.5.0